Second Version of Calculator

+++ Calculator/Models/NumberViewModel.cs
@@ -3,8 +3,8 @@ namespace Calculator.Models
 {
     public class NumberViewModel
     {
-        public double Number1 { get; set; }
-        public double Number2 { get; set; }
+        public decimal Number1 { get; set; }
+        public decimal Number2 { get; set; }
 
     }
 }
